+++ dev/main-react16/src/pages/vue2/vue2.js
@@ -83,7 +83,7 @@ function Vue2 () {
         // keep-alive
         // shadowDOM
         // destroy
-        // inline
+        inline
         // disable-scopecss
         // disableSandbox
         // disable-memory-router