+++ docs/zh-cn/changelog.md
@@ -10,11 +10,13 @@
 
 ### 1.0.0-rc.2
 
-`2023-10-26`
+`2023-10-30`
 
 - **New**
 
   - 🆕 新增了子应用全局变量`__MICRO_APP_STATE__`，用于标记当前应用的状态。
+  - 🆕 新增了子应用document变量`__MICRO_APP_NAME__`，用于标记document所属应用。
+  - 🆕 重写了原型方法`Node.prototype.parentNode`，用于处理特殊元素parentNode的指向问题。
 
 - **Bug Fix**
 
@@ -22,7 +24,7 @@
   - 🐞 修复了在iframe沙箱下开启`inline`模式导致通过`getElementsByTagName`获取script元素失败的问题。
 
 - **Update**
-
+  - 🆕 优化了内存占用，在iframe沙箱模式下默认开启inline模式。
   - 🚀 更新了官网文档。